+++ AdNetworkSupport/Facebook/FacebookNativeAdAdapter.m
@@ -9,8 +9,9 @@
 #import "FacebookNativeAdAdapter.h"
 #import "MPNativeAdConstants.h"
 #import "MPNativeAdError.h"
+#import "MPLogging.h"
 
-@interface FacebookNativeAdAdapter ()
+@interface FacebookNativeAdAdapter () <FBNativeAdDelegate>
 
 @property (nonatomic, readonly, strong) FBNativeAd *fbNativeAd;
 
@@ -24,6 +25,7 @@
 {
     if (self = [super init]) {
         _fbNativeAd = fbNativeAd;
+        _fbNativeAd.delegate = self;
 
         NSNumber *starRating = nil;
 
@@ -87,36 +89,42 @@
     return nil;
 }
 
-- (void)displayContentForURL:(NSURL *)URL rootViewController:(UIViewController *)controller
-                  completion:(void (^)(BOOL success, NSError *error))completionBlock
+- (BOOL)enableThirdPartyImpressionTracking
 {
-    NSError *error = nil;
-
-    if (!controller) {
-        NSDictionary *userInfo = [NSDictionary dictionaryWithObject:@"Cannot display content without a root view controller."
-                                                             forKey:MPNativeAdErrorContentDisplayErrorReasonKey];
-        error = [NSError errorWithDomain:MoPubNativeAdsSDKDomain
-                                    code:MPNativeAdErrorContentDisplayError
-                                userInfo:userInfo];
-
-        if (completionBlock) {
-            completionBlock(NO, error);
-        }
-
-        return;
-    }
-
-    [self.fbNativeAd handleClickWithViewController:controller
-                                        callback:^(FBNativeAd *nativeAd) {
-                                            if (completionBlock) {
-                                                completionBlock(YES, nil);
-                                            }
-                                        }];
+    return YES;
 }
 
-- (void)trackImpression
+- (BOOL)enableThirdPartyClickTracking
 {
-    [self.fbNativeAd logImpression];
+    return YES;
+}
+
+- (void)willAttachToView:(UIView *)view
+{
+    [self.fbNativeAd registerViewForInteraction:view withViewController:[self.delegate viewControllerForPresentingModalView]];
+}
+- (void)didDetachFromView:(UIView *)view
+{
+    [self.fbNativeAd unregisterView];
+}
+#pragma mark - FBNativeAdDelegate
+
+- (void)nativeAdWillLogImpression:(FBNativeAd *)nativeAd
+{
+    if ([self.delegate respondsToSelector:@selector(nativeAdWillLogImpression:)]) {
+        [self.delegate nativeAdWillLogImpression:self];
+    } else {
+        MPLogWarn(@"Delegate does not implement impression tracking callback. Impressions likely not being tracked.");
+    }
+}
+
+- (void)nativeAdDidClick:(FBNativeAd *)nativeAd
+{
+    if ([self.delegate respondsToSelector:@selector(nativeAdDidClick:)]) {
+        [self.delegate nativeAdDidClick:self];
+    } else {
+        MPLogWarn(@"Delegate does not implement click tracking callback. Clicks likely not being tracked.");
+    }
 }
 
 @end