+++ AdNetworkSupport/Facebook/FacebookNativeCustomEvent.h
@@ -8,7 +8,7 @@
 #import "MPNativeCustomEvent.h"
 
 /**
- * Certified with the Facebook iOS SDK version 3.14.1
+ * Certified with the Facebook iOS SDK version 3.18.2
  */
 
 @interface FacebookNativeCustomEvent : MPNativeCustomEvent