+++ Specs/MoPubSampleAppSpecs/SampleAppKIF/Extensions/KIFMPInstanceProvider.m
@@ -40,22 +40,22 @@ static KIFMPInstanceProvider *sharedProvider = nil;
 
 - (GSFullscreenAd *)buildGSFullscreenAdWithDelegate:(id<GSAdDelegate>)delegate GUID:(NSString *)GUID
 {
-    return [super buildGSFullscreenAdWithDelegate:delegate GUID:@"1d73efc1-c8c5-44e6-9b02-b6dd29374c1c"];
+    return [super buildGSFullscreenAdWithDelegate:delegate GUID:@"31d51c95-d79b-48c1-925e-ad328eb48c87"];
 }
 
 - (GSBannerAdView *)buildGreystripeBannerAdViewWithDelegate:(id<GSAdDelegate>)delegate GUID:(NSString *)GUID size:(CGSize)size
 {
-    return [super buildGreystripeBannerAdViewWithDelegate:delegate GUID:@"1d73efc1-c8c5-44e6-9b02-b6dd29374c1c" size:size];
+    return [super buildGreystripeBannerAdViewWithDelegate:delegate GUID:@"31d51c95-d79b-48c1-925e-ad328eb48c87" size:size];
 }
 
 - (IMInterstitial *)buildIMInterstitialWithDelegate:(id<IMInterstitialDelegate>)delegate appId:(NSString *)appId
 {
-    return [super buildIMInterstitialWithDelegate:delegate appId:@"eee4b638927c481cb5053469e209f631"];
+    return [super buildIMInterstitialWithDelegate:delegate appId:@"c8e9d75780cd439cad91d5def5200d25"];
 }
 
 - (IMBanner *)buildIMBannerWithFrame:(CGRect)frame appId:(NSString *)appId adSize:(int)adSize
 {
-    return [super buildIMBannerWithFrame:frame appId:@"eee4b638927c481cb5053469e209f631" adSize:adSize];
+    return [super buildIMBannerWithFrame:frame appId:@"c8e9d75780cd439cad91d5def5200d25" adSize:adSize];
 }
 
 @end